# docs/swagger.yaml
definitions:
  models.Class:
    type: object
  models.Student:
    type: object
info:
  contact: {}
paths:
  /classes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Class'
            type: array
      summary: Get all classes
      tags:
      - classes
    post:
      consumes:
      - application/json
      description: Create a new class
      parameters:
      - description: Class Data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Class'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Class'
      summary: Create a class
      tags:
      - classes
  /classes/{id}:
    delete:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
      summary: Delete a class
      tags:
      - classes
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
      summary: Get a class by ID
      tags:
      - classes
    put:
      consumes:
      - application/json
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      - description: Class Data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Class'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
      summary: Update a class
      tags:
      - classes
  /students:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
      summary: Get all students
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create a new student in a specific class
      parameters:
      - description: Student Data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
      summary: Create a student
      tags:
      - students
  /students/{id}:
    delete:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Delete a student
      tags:
      - students
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
      summary: Get a student by ID
      tags:
      - students
    put:
      consumes:
      - application/json
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student Data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
      summary: Update a student
      tags:
      - students
swagger: "2.0"